Fixed: the "hide toolbar" command did not work correctly under Linux. Fixed a problem with Linux compilation.
git-svn-id: https://vvvapp.svn.sourceforge.net/svnroot/vvvapp/trunk@68 b0e3558a-1426-0410-83de-bf875247a1ac

+++ src/data_interface/fb_volumes.cpp
@@ -62,7 +62,7 @@ void CVolumes::FB_DbDelete(void)
 
 	try {
 		st->Prepare( "EXECUTE PROCEDURE SP_DELETE_VOLUME( ? )" );
-		st->Set( 1, VolumeID );
+		st->Set( 1, (int32_t) VolumeID );
 		st->Execute();
 	}
 	catch( IBPP::SQLException& e ) {
